Adapt for use of internal catalog.  (See http://new.zope.org/Members/klm/ColDev/25 for explanation of new internal catalog.)
._collector_path: We need a stashed path to the containing collector,
for indexing happening outside acquisition.  (This will be handy if we
ever get to distributed collectors, where the issues can reside
outside the collector.)

.__init__(): Lots of shuffling, in my thrashing to make sure
order-sensitive activities happen ok.  Also, normalize
security_related setting to 0 or 1.

._set_collector_path(): Stash the path.

.edit(): Add explicit .reindexObject().  (I am getting ineffective -
not sure why this wasn't necessary before!  Urgh.)  Normalize
security_related to 0 or 1.

.do_action(): Add explicit .reindexObject().  (Same question arises as
for .edit().)

.upload_number(): For cataloging.

.indexObject(), .unindexObject(), .reindexObject(): Catalog against
internal catalog in addition to the portal one.

._get_internal_catalog(): New.

.manage_afterAdd(): do a ._set_collector_path() so the path tracks
collector moves and renames.

.Subject(): Radically simplify, to just the trait values as keywords.

.len(): Count the uploads.