Adding email address option for anonymous submitters (and alternate delivery address, for members).
When the submitter is anonymous, the explicitly specified email
address is essential - in its absence, supporters doing followups are
informed that the submitter will not get email notification.

.__init__(), .edit(), ._send_update_notice(): When the submitter is a
member (with an email address preference), an explicitly specified
email address overrides the preference.  If the email address is the
*same* as the preference, then the explicitly specified setting is
wiped (set to None), and the preference is used from then on.  This
way, the submitter's preference is tracked thereafter.

.no_submitter_email(): Helper for followup form, so the supporter can
be informed that the submitter won't be getting a notification.